requirements.txt
matplotlib
numpy
opencv-python
pdbpp
scipy
Pillow
gdown
protobuf==3.20.0
chardet
pytype
types-requests
dynaconf